cookiecutter_mlops/requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
alabaster==0.7.12
antlr4-python3-runtime==4.8
appdirs==1.4.4
arrow==1.2.1
async-timeout==4.0.2
asyncssh==2.8.1
atpublic==3.0.1
attrs==21.4.0
awscli==1.22.33
Babel==2.9.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
binaryornot==0.4.4
botocore==1.23.33
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
commonmark==0.9.1
configobj==5.0.6
configparser==5.2.0
cookiecutter==1.7.3
coverage==6.2
cryptography==36.0.1
cycler==0.11.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.6.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
dpath==2.0.5
dulwich==0.20.30
dvc==2.9.3
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
flake8==4.0.1
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.28.5
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
google-api-core==2.4.0
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
grandalf==0.6
helper==2.5.0
httplib2==0.20.2
hydra-core==1.1.1
idna==3.3
imagesize==1.3.0
iniconfig==1.1.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1641325328741/work/dist/ipykernel-6.6.1-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1638470227058/work
isort==5.10.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1637175083648/work
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1636814260563/work
kiwisolver==1.3.2
mailchecker==4.1.8
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mccabe==0.6.1
multidict==5.2.0
nanotime==0.5.2
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
networkx==2.6.3
numpy==1.21.0
oauth2client==4.1.3
omegaconf==2.1.1
packaging==21.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.9.0
pathtools==0.1.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
phonenumbers==8.12.41
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.0.0
pipreqs==0.4.11
pluggy==1.0.0
ply==3.11
poyo==0.5.0
promise==2.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1639065841292/work
protobuf==3.19.2
psutil==5.9.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
PyDrive2==1.10.0
pyflakes==2.4.0
pygit2==1.7.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1640969120816/work
pygtrie==2.4.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
python-benedict==0.24.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==0.19.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==19.0.2
requests==2.27.0
rich==11.0.0
rsa==4.7.2
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scmrepo==0.0.7
sentry-sdk==1.5.2
shortuuid==1.0.8
shtab==1.5.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
-e git+https://github.com/RianLeevinson/dtu_mlops_exercises.git@454e002752d616517fcc21091ece4e5e5e5398ec#egg=src&subdirectory=cookiecutter_mlops
subprocess32==3.5.4
tabulate==0.8.9
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
torch==1.10.1
torchserve==0.5.2
torchvision==0.11.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094708661/work
tqdm==4.62.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
voluptuous==0.12.2
wandb==0.12.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
xmltodict==0.12.0
yarg==0.1.9
yarl==1.7.2
yaspin==2.1.0
zc.lockfile==2.0
